# attached_assets/openapi_1776713333572.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gmail/auth:
    get:
      operationId: startGmailAuth
      tags: [doctrine]
      summary: Get Google OAuth URL to connect a Gmail account
      parameters:
        - name: x-api-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
                required:
                  - authUrl
  /gmail/accounts:
    get:
      operationId: getGmailAccounts
      tags: [doctrine]
      summary: List connected Gmail accounts
      parameters:
        - name: x-api-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connected accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailAccountsResponse"
  /gmail/accounts/{id}:
    delete:
      operationId: disconnectGmailAccount
      tags: [doctrine]
      summary: Disconnect a Gmail account
      parameters:
        - name: x-api-key
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnect result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisconnectResult"
  /gmail/accounts/{id}/settings:
    put:
      operationId: updateAccountSettings
      tags: [doctrine]
      summary: Update follow-up settings for an account
      parameters:
        - name: x-api-key
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateSettingsResult"
...
components:
  schemas:
...
    StageTiming:
      type: object
      properties:
        minDays:
          type: integer
        maxDays:
          type: integer
      required:
        - minDays
        - maxDays
    GmailAccount:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        isConnected:
          type: boolean
        stageTiming:
          type: array
          items:
            $ref: "#/components/schemas/StageTiming"
        sendDays:
          type: array
          items:
            type: integer
        sendHourStart:
          type: integer
        sendHourEnd:
          type: integer
        maxFollowups:
          type: integer
        doctrineLabel:
          type: string
        testMode:
          type: boolean
        requireApproval:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - isConnected
        - stageTiming
        - sendDays
        - sendHourStart
        - sendHourEnd
        - maxFollowups
        - doctrineLabel
        - testMode
        - requireApproval
    GmailAccountsResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/GmailAccount"
      required:
        - accounts
    DisconnectResult:
      type: object
      properties:
        success:
          type: boolean
        email:
          type: string
      required:
        - success
        - email
    AccountSettingsUpdate:
      type: object
      properties:
        stageTiming:
          type: array
          items:
            $ref: "#/components/schemas/StageTiming"
        sendDays:
          type: array
          items:
            type: integer
        sendHourStart:
          type: integer
        sendHourEnd:
          type: integer
        maxFollowups:
          type: integer
        doctrineLabel:
          type: string
        testMode:
          type: boolean
        requireApproval:
          type: boolean
    UpdateSettingsResult:
      type: object
      properties:
        success:
          type: boolean
        account:
          $ref: "#/components/schemas/GmailAccount"
      required:
        - success
...